Multi Module 샘플 코드 클론  - 에러 수정 Rename
