+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,12 @@
         </activity>
 
         <activity android:name=".views.login.LoginActivity"
+            android:screenOrientation="portrait"
             android:windowSoftInputMode="stateHidden|adjustPan"/>
+
+        <activity android:name=".views.search.MovieSearchActivity"
+            android:screenOrientation="portrait"/>
+
     </application>
 
 </manifest>
