+++ domain/src/main/java/com/kim344/domain/usecase/GetMoviesUseCase.kt
@@ -1,12 +1,13 @@
 package com.kim344.domain.usecase
 
-import com.kim344.domain.model.search.Movie
+import com.kim344.domain.search.Movie
 import com.kim344.domain.repository.MovieRepository
 import io.reactivex.Flowable
 import javax.inject.Inject
 
-class GetMoviesUseCase @Inject constructor(private val repository: MovieRepository){
+class GetMoviesUseCase @Inject constructor(private val repository: MovieRepository) {
     fun execute(
         query: String
     ): Flowable<List<Movie>> = repository.getSearchMovies(query)
+
 }
