+++ domain/src/main/java/com/kim344/domain/usecase/GetPagingMoviesUseCase.kt
@@ -1,11 +1,11 @@
 package com.kim344.domain.usecase
 
-import com.kim344.domain.model.search.Movie
+import com.kim344.domain.search.Movie
 import com.kim344.domain.repository.MovieRepository
 import io.reactivex.Single
 import javax.inject.Inject
 
-class GetPagingMovieUseCase @Inject constructor(private val repository: MovieRepository) {
+class GetPagingMoviesUseCase @Inject constructor(private val repository: MovieRepository) {
     fun execute(
         query: String,
         offset: Int
